refactor: some changes into naming of command functions
