+++ fileManager/commandsManager/commands/ls.js
@@ -1,9 +1,9 @@
-import { checkAccess } from "../checkFileAccess.js";
+import { checkAccess } from "../supportiveFileFuncs.js";
 import { getSetting } from "../../settings.js";
 import { readdir } from "fs/promises";
 import { constants } from "fs";
 
-export async function ls() {
+export async function ls(nonProcessedInput) {
     const path = await checkAccess(getSetting("location"), constants.R_OK);
     const dirContent = await readdir(path);
     for await (let file of dirContent) {
