+++ fileManager/commandsManager/commands/mv.js
@@ -1,4 +1,4 @@
-import { checkAccess, makePathAbsolute } from "../checkFileAccess.js";
+import { checkAccess, makePathAbsolute } from "../supportiveFileFuncs.js";
 import { copyFile, rm } from "fs/promises";
 
 export async function mv(nonProcessedInput) {
